screenOps/requirement.txt
numpy==1.21.2
paho-mqtt==1.5.1
Pillow==8.3.2
pkg_resources==0.0.0
RPi.GPIO==0.7.0
spidev==3.5
